Split code into scripts
