+++ prueba.py
@@ -1,7 +1,7 @@
-import tree
+from src.genetree import Genetree
 import pandas as pd
 
 data = pd.read_csv("iris-species/Iris.csv")
-tree = tree.Genetreec(data[['SepalLengthCm', 'SepalWidthCm', 'PetalLengthCm', 'PetalWidthCm']], data[['Species']])
+tree = Genetree(data[['SepalLengthCm', 'SepalWidthCm', 'PetalLengthCm', 'PetalWidthCm']], data[['Species']])
 tree.warm()
 tree.root.plot()
